Turn int to float

+++ main.py
@@ -6,8 +6,8 @@ goodnesses = []
 with open('ratings.txt') as ratings:
   for line in ratings:
     pair = line.split(' ')
-    lawfulnesses.append(int(pair[0]))
-    goodnesses.append(int(pair[1]))
+    lawfulnesses.append(float(pair[0]))
+    goodnesses.append(float(pair[1]))
     
 plt.xlabel('Lawfulness')
 plt.ylabel('Goodness')
